Moved the main() methods of the tokenizer and syntax tree generator to a separate file, to allow future files to link to them.
