+++ README.md
@@ -7,18 +7,17 @@ This is a personal project, written in C, to practice skills by writing an inter
 The language will be very simple to make it easy to implement. Currently it will implement the following functions:
  - Addition (+ a b c ...) - adds all variables a, b, c ...
  - Subtraction (- a b c ...) - subtracts the value of all variables after the first from the first variable. Equivalent to (- a (+ b c ...))
- - 
-# Examples
-
 
 # Compiling source files
 I prefer to use [Microsoft Visual Studio](https://www.visualstudio.com/en-us/visual-studio-homepage-vs.aspx)'s [command line C compiler, cl](https://msdn.microsoft.com/en-us/library/f35ctcxw.aspx). Any C compiler that builds to the C runtime environment should do the trick.
 
-- To compile the tokenizer, compile [tokenizer.c](src/tokenizer.c)
-- To compile the standalone parser, compile [parser.c](src/main/parser.c)
-- To compile the syntax tree generator (which includes the parser), compile [syntax.c](src/syntax.c)
+Source files for the final executables are in [src/main](src/main/). Files in [src](src/) are all helper functions for those other files.
 
-For example, `cl src/tokenizer.c` would compile the tokenizer and create `tokenizer.exe` in the current directory.
+- To compile the tokenizer, compile [tokenizer.c](src/main/tokenizer.c)
+- To compile the standalone parser, compile [parser.c](src/main/parser.c)
+- To compile the syntax tree generator (which includes the parser), compile [syntax.c](src/main/syntax.c)
+
+For example, `cl src/main/tokenizer.c` would compile the tokenizer and create `tokenizer.exe` in the current directory.
 
 # Using the executable interpreter components
 (Read more at [examples](examples/)
