+++ examples/parser_test.jtoken
@@ -1,5 +1,5 @@
-1 1 (
-3 1 1
-3 1 2
-3 1 3
-2 1 )
+1 1 1 (
+3 1 3 1
+3 1 5 2
+3 1 7 3
+2 1 9 )
